feat: implement MainMenu component for improved navigation structure
